Implement the random point on key method, based on the bias of previous key.
Added a method to pick a random point on the closest compass point direction - sub-square of the key to press.
Also added a method to deal with a random point to a single key.

This commit has some messy code-containing-comments, which should be filtered out after we are comfortable with the testing.

Signed-off-by: CalumD <github@clumd.com>