Add support for writing to files
The program can now take multiple words and output them to
individual files. The files are output in the directory specified
with the -o command-line option, and are named 0.svg, 1.svg,
2.svg, and so on.

The -i option is still not supported; currently the program
recieves words to render from its command-line arguments.

The -o option is required currently, though ideally the program
would print to stdout if it is not given an output file/directory.

The SVGs under construction are now held in a buffer before being
written to file all-at-once, rather than output element-at-a-time
like previously. The buffers will automatically expand when
necessary.

The svg_-prefixed functions could possibly be moved to their own
utility file.

The svg_-prefixed functions could possibly be moved to their own
utility file.