Add support for reading input from file
The `-i` option can be used to specify a file to take input from.
Words added to the command line are processed after those found in
the input file.

Passing `-` as the input file name instead reads from stdin; reading
from a file literally called `-` can be done by addressing it as
`./-`.